refactor
Migrating SceneEvent and SceneEventType into their own files.
Removing virtual from the ClientLoadSceneCoroutine and ClientUnloadSceneCoroutine.
