+++ com.unity.netcode.gameobjects/Runtime/SceneManagement/SceneEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c940af5c809c3bc4d979111290f30156
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
