+++ com.unity.netcode.gameobjects/Runtime/SceneManagement/SceneEventType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2e7113e167129c243a7aafd9e9fc7236
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
